Refactored how app listens and pre-caches points/times
