+++ src/ui/setup/index.ts
@@ -25,7 +25,18 @@ import { ThreadFront } from "protocol/thread";
 import { replayClient } from "shared/client/ReplayClientContext";
 import { ReplayClient } from "shared/client/ReplayClient";
 import { getPreferredLocation } from "ui/utils/preferredLocation";
-import { Location } from "@replayio/protocol";
+import {
+  ExecutionPoint,
+  loadedRegions as LoadedRegions,
+  Location,
+  PauseData,
+  PauseId,
+  TimeStampedPointRange,
+} from "@replayio/protocol";
+import { addPauseDataListener } from "protocol/thread/pause";
+import { preCacheObjects } from "bvaughn-architecture-demo/src/suspense/ObjectPreviews";
+import { trackExecutionPointPauseIds } from "bvaughn-architecture-demo/src/suspense/PauseCache";
+import { preCacheExecutionPointForTime } from "@bvaughn/src/suspense/PointsCache";
 
 declare global {
   interface Window {
@@ -107,6 +118,38 @@ export async function bootstrapApp() {
   if (typeof window === "undefined") {
     return store;
   }
+
+  // Connect data in legacy Redux stores (like PauseData or points) to the newer Suspense caches.
+  // This avoids requiring the new components from requesting redundant data.
+  // In the case of PauseData, it's extra important– because the backend won't re-send the same data twice no matter how many times you ask for it.
+  addPauseDataListener((pauseId: PauseId, executionPoint: ExecutionPoint, pauseData: PauseData) => {
+    const { objects } = pauseData;
+    if (objects) {
+      // Be sure to clone object data before pre-caching it.
+      // Otherwise ValueFronts might deeply mutate it and change its structure.
+      const cloned = JSON.parse(JSON.stringify(objects));
+      preCacheObjects(pauseId, cloned);
+    }
+
+    trackExecutionPointPauseIds(executionPoint, pauseId);
+  });
+
+  // Listen for changes in loaded regions and pre-caches the points.
+  // eslint-disable-next-line @typescript-eslint/no-floating-promises
+  ThreadFront.listenForLoadChanges((loadedRegions: LoadedRegions) => {
+    loadedRegions.indexed.forEach(({ begin, end }) => {
+      preCacheExecutionPointForTime(begin);
+      preCacheExecutionPointForTime(end);
+    });
+    loadedRegions.loaded.forEach(({ begin, end }) => {
+      preCacheExecutionPointForTime(begin);
+      preCacheExecutionPointForTime(end);
+    });
+    loadedRegions.loading.forEach(({ begin, end }) => {
+      preCacheExecutionPointForTime(begin);
+      preCacheExecutionPointForTime(end);
+    });
+  });
 
   // Wire up new Console and Object Inspector to the Redux logic for preferred source.
   (replayClient as ReplayClient).injectGetPreferredLocation(
